server/requirements.txt
annotated-types==0.5.0
blinker==1.6.2
click==8.1.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.7
pydantic==2.2.0
pydantic_core==2.6.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.6
